refactor files table for clarity
